refactor: move method position (raceOneTime to car, raceOneStage to race)
