+++ src/main/java/racingcar/controller/RacingGame.java
@@ -1,9 +1,6 @@
 package racingcar.controller;
 
-import java.util.List;
-import racingcar.domain.Car;
 import racingcar.domain.Race;
-import racingcar.domain.RandomNumberGenerator;
 import racingcar.domain.Times;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
@@ -22,17 +19,17 @@ public class RacingGame {
 
     private static void repeatGame(Times times, Race race) {
         while (times.isOngoing()) {
-            raceOneStage(race);
+            race.raceOneStage();
             times.decrease();
         }
     }
 
-    public static void raceOneStage(Race race) {
-        List<Car> cars = race.getCars();
-        for (Car car : cars) {
-            race.raceOneTime(car, RandomNumberGenerator.getRandomNumber());
-        }
-        OutputView.printCurrentPosition(cars);
-        System.out.println();
-    }
+//    public static void raceOneStage(Race race) {
+//        List<Car> cars = race.getCars();
+//        for (Car car : cars) {
+//            race.raceOneTime(car, RandomNumberGenerator.getRandomNumber());
+//        }
+//        OutputView.printCurrentPosition(cars);
+//        System.out.println();
+//    }
 }
