+++ src/main/java/racingcar/domain/Race.java
@@ -2,20 +2,22 @@ package racingcar.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import racingcar.view.OutputView;
 
 public class Race {
 
-    private static final int MIN_NUMBER_FOR_START = 4;
     private final List<Car> cars;
 
     public Race(List<Car> cars) {
         this.cars = cars;
     }
 
-    public void raceOneTime(Car car, int condition) {
-        if (condition >= MIN_NUMBER_FOR_START) {
-            car.go();
+    public void raceOneStage() {
+        for (Car car : cars) {
+            car.raceOneTime(RandomNumberGenerator.getRandomNumber());
         }
+        OutputView.printCurrentPosition(cars);
+        System.out.println();
     }
 
     public List<String> getWinnerNames() {
@@ -35,9 +37,5 @@ public class Race {
             maximum = car.comparePosition(maximum);
         }
         return maximum;
-    }
-
-    public List<Car> getCars() {
-        return cars;
     }
 }
