+++ src/test/java/racingcar/domain/RaceTest.java
@@ -1,47 +1,5 @@
 package racingcar.domain;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-import java.util.ArrayList;
-import java.util.List;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
 class RaceTest {
 
-    @DisplayName("난수가 4일때 배열의 숫자가 증가하는지 확인한다")
-    @Test
-    public void checkWhenRandomNumberIsFour() {
-        List<Car> cars = new ArrayList<>();
-        cars.add(new Car("audi"));
-        cars.add(new Car("benz"));
-        cars.add(new Car("bmw"));
-        Race race = new Race(cars);
-        race.raceOneTime(cars.get(1), 4);
-        assertThat(cars.get(1).getPosition()).isEqualTo(1);
-    }
-
-    @DisplayName("난수가 5일때 배열의 숫자가 증가하는지 확인한다")
-    @Test
-    public void checkWhenRandomNumberIsFive() {
-        List<Car> cars = new ArrayList<>();
-        cars.add(new Car("audi"));
-        cars.add(new Car("benz"));
-        cars.add(new Car("bmw"));
-        Race race = new Race(cars);
-        race.raceOneTime(cars.get(1), 5);
-        assertThat(cars.get(1).getPosition()).isEqualTo(1);
-    }
-
-    @DisplayName("난수가 3일때 배열의 숫자가 증가하는지 확인한다")
-    @Test
-    public void checkWhenRandomNumberIsThree() {
-        List<Car> cars = new ArrayList<>();
-        cars.add(new Car("audi"));
-        cars.add(new Car("benz"));
-        cars.add(new Car("bmw"));
-        Race race = new Race(cars);
-        race.raceOneTime(cars.get(1), 3);
-        assertThat(cars.get(1).getPosition()).isEqualTo(0);
-    }
 }
